crawler-vk/pyproject.toml
[project]
name = "crawler-vk"
version = "0.1.0"
description = ""
authors = [{ name = "Ivan Baklanov", email = "nymless.ib@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
packages = [{ include = "src" }]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
mypy = "^1.15.0"
types-requests = "^2.32.0.20241016"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["E501"]
